fixed error in articlereview - made auth-context a typescript
